// src/app/admin/shared/models/public-pricing.model.ts
export class PublicPricing {
    egyptianPrice: number;
    egyptianPriceNote: string;
    foreignerPrice: number;
    foreignerPriceNote: string;
    studentForeignerPrice: number;
    studentForeignerPriceNote: string;
    groupOfOnePrice:number;
    groupOfOneNote:string;
    groupOfMorePrice:number;
    groupOfMoreNote:string;
}